Bind du modèle sur le GUI

+++ Pimp-my-fridge-java/src/fr/epickiwi/pmf/model/FridgeSettings.java
@@ -1,35 +1,32 @@
 package fr.epickiwi.pmf.model;
 
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.FloatProperty;
-import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleFloatProperty;
+import javafx.beans.property.*;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 
 public class FridgeSettings {
 
-    private FloatProperty orderTemperature;
+    private DoubleProperty orderTemperature;
     private BooleanProperty avoidCondensation;
     private ObservableList<TemperaturePreset> presets;
 
     public FridgeSettings() {
-        this.orderTemperature = new SimpleFloatProperty(0);
+        this.orderTemperature = new SimpleDoubleProperty(0);
         this.avoidCondensation = new SimpleBooleanProperty(false);
         this.presets = FXCollections.observableArrayList();
     }
 
     /* ----- GETTERS AND SETTERS ----- */
 
-    public float getOrderTemperature() {
+    public double getOrderTemperature() {
         return orderTemperature.get();
     }
 
-    public FloatProperty orderTemperatureProperty() {
+    public DoubleProperty orderTemperatureProperty() {
         return orderTemperature;
     }
 
-    public void setOrderTemperature(float orderTemperature) {
+    public void setOrderTemperature(double orderTemperature) {
         this.orderTemperature.set(orderTemperature);
     }
